+++ examples/typed_access/codegen_demo/__init__.py
@@ -1,0 +1,1 @@
+"""Type-safe Smalltalk code-generation demo."""